pyproject.toml
[project]
name = "toya"
version = "1.0.1"
description = "Configuration management with deferred binding and value templating"
authors = [
    { name = "Sergei Galushkin", email = "cs0ip@ya.ru" }
]
dependencies = [
    "pyyaml>=6.0.2",
    "mergedeep>=1.3.4",
    "jinja2>=3.1.6",
]
readme = "README.md"
requires-python = ">= 3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Topic :: File Formats",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed"
]
license = "MIT"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.4.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/toya"]

[tool.pytest.ini_options]
testpaths = ["tests"]
